Move generation of recent patient UI list items to a factory method
